+++ internal/delivery/httpService/routes.go
@@ -14,9 +14,10 @@ func RegisterRoutes(
 	sessionHandler handlers.SessionHandler,
 	walletHandler handlers.WalletHandler,
 	computerHandler handlers.ComputerHandler,
+	authHandler handlers.AuthHandler,
 ) {
-	r.Post("/register", userHandler.RegisterUser)
-	r.Post("/login", userHandler.LoginUser)
+	r.Post("/auth/register", authHandler.Register)
+	r.Post("/auth/login", authHandler.Login)
 
 	r.Get("/tariff", tariffHandler.GetTariff)
 	r.Get("/tariff/{id}", tariffHandler.GetTariffByID)
@@ -24,16 +25,20 @@ func RegisterRoutes(
 	r.Group(func(protected chi.Router) {
 		protected.Use(middleware.AuthMiddleware)
 
-		protected.Get("/info", userHandler.InfoUser)
 		//protected.Post("/tariff", tariffHandler.AddTariff)
 		//protected.Put("/tariff/{id}", tariffHandler.ChangeTariff)
 		//protected.Delete("/tariff/{id}", tariffHandler.DeleteTariff)
-		protected.Get("/sessions/active", sessionHandler.GetActiveSessions)
+
 		protected.Get("/users", userHandler.GetUsers)
 		protected.Get("/users/{id}", userHandler.GetUserByID)
-		protected.Post("/session/start", sessionHandler.StartSession)
-		protected.Post("/session/end", sessionHandler.EndSession)
+		protected.Get("/users/info", userHandler.InfoUser)
+
 		protected.Put("/pay", walletHandler.PutMoneyOnWallet)
+
+		protected.Post("/session/start", sessionHandler.StartSession)
+		protected.Get("/session/active", sessionHandler.GetActiveSessions)
+		protected.Post("/session/end", sessionHandler.EndSession)
+
 		protected.Get("/computers", computerHandler.GetComputersStatus)
 		protected.Post("/computers", computerHandler.AddComputer)
 		protected.Delete("/computers/{id}", computerHandler.DeleteComputer)
